Big commit changing to ChipPiles for internal use in Player and Pot.
Table came along for the ride for stuff like interacting with Pot's skim stuff.

Testing all this.

The point here is that next I plan to make all transfers from ChipPile to ChipPile done via a ChipConduit, after which there will be no more subsequent lines
of adding to one and subtracting from another.  This will allow me to publish via pubsub all of these transactions so brains can watch them and so I can report
on them.

This is all feeling a bit ugly so who knows if it'll stick around.